requirements.txt
celery==5.2.7
Django==4.1
djangorestframework==3.13.1
djangorestframework-gis==1.0
djangorestframework-simplejwt==5.2.0
psycopg2-binary==2.9.3
